backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.32
pydantic-settings
python-dotenv
psycopg2-binary
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-multipart
